perf: code optimization for portkeyWallet class
